+++ src/equation/cg.cpp
@@ -71,7 +71,7 @@ int equation::CG<T>::monolish_CG(matrix::CRS<T> &A, vector<T> &x,
       return MONOLISH_SOLVER_SUCCESS;
     }
 
-    if ( std::isnan(resid)) {
+    if (std::isnan(resid)) {
       return MONOLISH_SOLVER_RESIDUAL_NAN;
     }
   }