+++ include/monolish_vml_float.hpp
@@ -587,8 +587,7 @@ void div(const matrix::CRS<float> &A, const matrix::CRS<float> &B,
  * - GPU acceleration: true
  *    - # of data transfer: 0
  */
-void add(const matrix::CRS<float> &A, const float alpha,
-         matrix::CRS<float> &C);
+void add(const matrix::CRS<float> &A, const float alpha, matrix::CRS<float> &C);
 
 /**
  * @brief single precision scalar and Dence matrix subtraction (C[i][j] =
@@ -602,8 +601,7 @@ void add(const matrix::CRS<float> &A, const float alpha,
  * - GPU acceleration: true
  *    - # of data transfer: 0
  */
-void sub(const matrix::CRS<float> &A, const float alpha,
-         matrix::CRS<float> &C);
+void sub(const matrix::CRS<float> &A, const float alpha, matrix::CRS<float> &C);
 
 /**
  * @brief single precision scalar and Dence matrix multiplication (C[i][j] =
@@ -617,8 +615,7 @@ void sub(const matrix::CRS<float> &A, const float alpha,
  * - GPU acceleration: true
  *    - # of data transfer: 0
  */
-void mul(const matrix::CRS<float> &A, const float alpha,
-         matrix::CRS<float> &C);
+void mul(const matrix::CRS<float> &A, const float alpha, matrix::CRS<float> &C);
 
 /**
  * @brief single precision scalar and Dence matrix division (C[i][j] = A[i][j] +
@@ -632,8 +629,7 @@ void mul(const matrix::CRS<float> &A, const float alpha,
  * - GPU acceleration: true
  *    - # of data transfer: 0
  */
-void div(const matrix::CRS<float> &A, const float alpha,
-         matrix::CRS<float> &C);
+void div(const matrix::CRS<float> &A, const float alpha, matrix::CRS<float> &C);
 
 /**
  * @brief power to single precision CRS matrix elements by single precision
@@ -654,8 +650,7 @@ void pow(const matrix::CRS<float> &A, const matrix::CRS<float> &B,
  * - Multi-threading: true
  * - GPU acceleration: true
  **/
-void pow(const matrix::CRS<float> &A, const float alpha,
-         matrix::CRS<float> &C);
+void pow(const matrix::CRS<float> &A, const float alpha, matrix::CRS<float> &C);
 
 /**
  * @brief sqrt to single precision CRS matrix elements (y[0:N] = sqrt(a[0:N]))