+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Unreleased
 -----------
 
 ### Changed 
+- update allgebra 20.10.1 -> 20.12.1 https://gitlab.ritc.jp/ricos/monolish/-/terge_requests/173
 - CIのRunner指定をsmタグに変更 https://gitlab.ritc.jp/ricos/monolish/-/terge_requests/171
 
 ### Fixed